+++ com/test/SellingDiscount.java
@@ -2,7 +2,7 @@ package com.test;
 
 public class SellingDiscount implements Selling {
 
-	private final static long mongoPrice  = 20L;
+	private final static double mongoPrice  = 20L;
 	
 	private final static Selling  sell = new SellingDiscount();
 	
@@ -15,11 +15,11 @@ public class SellingDiscount implements Selling {
 	}
 	
 	@Override
-	public long totalSum(int kilo) throws Exception {
+	public double totalSum(int kilo) throws Exception {
 		if(kilo<0) {
 			throw new Exception("斤数小于0");
 		}
-		return (long) (kilo*applePrice+kilo*berryPrice*0.8+kilo*mongoPrice);
+		return   kilo*applePrice+kilo*berryPrice*0.8+kilo*mongoPrice;
 	}
 
 }
