+++ com/test/SellingReduce.java
@@ -2,7 +2,7 @@ package com.test;
 
 public class SellingReduce implements Selling {
 	
-	private final static long mongoPrice  = 20L;
+	private final static double mongoPrice  = 20;
 
 	private final static Selling  sell = new SellingReduce();
 	
@@ -16,12 +16,12 @@ public class SellingReduce implements Selling {
 
 	
 	@Override
-	public long totalSum(int kilo) throws Exception {
+	public double totalSum(int kilo) throws Exception {
 		if(kilo<0) {
 			throw new Exception("斤数小于0");
 		}
-		long beforeSum = kilo*applePrice+kilo*berryPrice+kilo*mongoPrice;
-		return beforeSum-beforeSum/100*10;
+		int beforeSum =  (int) (  kilo*applePrice+kilo*berryPrice+kilo*mongoPrice) ;
+		return (double)beforeSum-beforeSum/100*10;
 	}
 
 }
